# [FIX] stock: Add index of product_id field of stock.move.line model