Add index for field product_id off StockMoveLine model.
Resolve performance issues with big databases.

**Without index**:
```
EXPLAIN ANALYZE SELECT "stock_move_line".id FROM "stock_move_line" WHERE ("stock_move_line"."product_id" in (236)) ORDER BY "stock_move_line"."id"
----
Sort  (cost=5762.41..5762.42 rows=3 width=4) (actual time=36.188..36.190 rows=10 loops=1)
  Sort Key: id
  Sort Method: quicksort  Memory: 25kB
  ->  Seq Scan on stock_move_line  (cost=0.00..5762.39 rows=3 width=4) (actual time=27.923..36.181 rows=10 loops=1)
        Filter: (product_id = 236)
        Rows Removed by Filter: 80081
Planning Time: 0.137 ms
Execution Time: 36.209 ms
```

==================================================

**With index**:
```
EXPLAIN ANALYZE SELECT "stock_move_line".id FROM "stock_move_line" WHERE ("stock_move_line"."product_id" in (237)) ORDER BY "stock_move_line"."id"
----
Sort  (cost=8.50..8.50 rows=3 width=4) (actual time=0.048..0.049 rows=10 loops=1)
  Sort Key: id
  Sort Method: quicksort  Memory: 25kB
  ->  Index Scan using stock_move_line_product_id_index on stock_move_line  (cost=0.42..8.47 rows=3 width=4) (actual time=0.035..0.040 rows=10 loops=1)
        Index Cond: (product_id = 237)
Planning Time: 0.363 ms
Execution Time: 0.063 ms
```

---
For example, it is needed in v.15 due to the field `product_stock_quant_ids` added in `stock_barcode`. This field triggers the above queries (without the `EXPLAIN ANALYZE` of course) in the method `_modified_triggers()` when validating a transfer with many stock move lines.
It results in an execution time as follow:
![image](https://user-images.githubusercontent.com/29302288/177111016-e48a9e18-e156-4dd9-9c2f-e1571b1f7249.png)

I discussed it with the Performance Team and what resulted in our discussion was that adding the index by default was justified.

---
OPW-2893131

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr